+++ nost_tools/__init__.py
@@ -1,11 +1,10 @@
-__version__ = "2.0.5"
+__version__ = "2.1.0"
 
 from .application import Application
 from .application_utils import ConnectionConfig, ModeStatusObserver, TimeStatusPublisher
 from .configuration import ConnectionConfig
 from .entity import Entity
-
-# from .logger_application import LoggerApplication
+from .logger_application import LoggerApplication
 from .managed_application import ManagedApplication
 from .manager import Manager, TimeScaleUpdate
 from .observer import Observable, Observer